↺ merge  svn merge -r 2551:2685 https://amethyst.yui.ne.jp/svn-dev/pc98/np2/trunk  svn merge -r 2686:2733 https://amethyst.yui.ne.jp/svn-dev/pc98/np2/trunk
git-svn-id: http://amethyst.yui.ne.jp/svn/pc98/np2/branches/yui@2734 7c2116fe-53d0-4aec-ab59-80607f05abc0

+++ reform_ui/embed/menu/dlgcfg.c
@@ -310,8 +310,8 @@ const OEMCHAR	*str;
 	else {
 		val = 22050;
 	}
-	if (np2cfg.samplingrate != (UINT16)val) {
-		np2cfg.samplingrate = (UINT16)val;
+	if (np2cfg.samplingrate != val) {
+		np2cfg.samplingrate = val;
 		update |= SYS_UPDATECFG | SYS_UPDATERATE;
 		soundrenewal = 1;
 	}
